requirements.txt
streamlit
PyMuPDF
sentence-transformers
faiss-cpu
openai
numpy
